# OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tutorial Management API",
    "description": "This API exposes endpoints to manage tutorials.",
    "termsOfService": "https://www.bezkoder.com/terms",
    "contact": {
      "name": "My Name",
      "url": "https://www.fct.unl.pt",
      "email": "myemail@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://choosealicense.com/licenses/mit/"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Server URL in Development environment"
    },
    {
      "url": "https://api.example.com/",
      "description": "Server URL in Production environment"
    }
  ],
  "paths": {
...
    "/apartments/{apartmentId}/period": {
      "post": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "createApartmentPeriod",
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Period added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PeriodInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid period date.",
            "content": {}
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Apartment not found.",
            "content": {}
          },
          "409": {
            "description": "Period already exists in this date.",
            "content": {}
          }
        }
      }
    },
    "/apartments/{apartmentId}/booking/{username}": {
      "post": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "bookApartment",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookApartmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Apartment was successfully booked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingInfoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid booking data; Apartment has no available periods.",
            "content": {}
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Client not found; Apartment not found",
            "content": {}
          },
          "409": {
            "description": "Booking is unavailable for this period.",
            "content": {}
          }
        }
      }
    },
...
    "/periods/{periodID}": {
      "get": {
        "tags": [
          "period-controller"
        ],
        "operationId": "getPeriod",
        "parameters": [
          {
            "name": "periodID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/owners/{ownerId}/apartments": {
      "get": {
        "tags": [
          "owner-controller"
        ],
        "operationId": "getOwnerApartments",
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner Apartments.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApartmentInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Owner not found.",
            "content": {}
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
...
    "/apartments/{apartmentId}": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "getApartmentById",
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Apartment info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApartmentInfoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/apartments/{apartmentId}/reviews": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "getApartmentReviews",
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Apartment's Reviews.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Apartment not found."
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/apartments/{apartmentId}/periods": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "getApartmentPeriods",
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of apartment periods",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PeriodInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Apartment not found.",
            "content": {}
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/apartments/{apartmentId}/periods/{periodId}": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "getApartmentPeriod",
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "periodId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of apartments"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/apartments/{apartmentId}/availablePeriods": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "getApartmentAvailablePeriods",
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of available apartment periods",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailablePeriodInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Apartment not found.",
            "content": {}
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/apartments/periods": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "getAvailableApartmentsForPeriod",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of apartments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApartmentInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid period data.",
            "content": {}
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/apartments/": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "getApartments",
        "responses": {
          "200": {
            "description": "List of apartments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApartmentInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ReviewInfoDTO": {
        "required": [
          "clientName"
        ],
        "type": "object",
        "properties": {
          "clientName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PeriodCreateDTO": {
        "required": [
          "endDate",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "PeriodInfoDTO": {
        "required": [
          "apartId",
          "available",
          "date"
        ],
        "type": "object",
        "properties": {
          "apartId": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "available": {
            "type": "boolean"
          }
        }
      },
      "BookApartmentDTO": {
        "required": [
          "checkIn",
          "checkOut",
          "guests"
        ],
        "type": "object",
        "properties": {
          "checkIn": {
            "type": "string",
            "format": "date"
          },
          "checkOut": {
            "type": "string",
            "format": "date"
          },
          "guests": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApartmentInfoDTO": {
        "required": [
          "amenities",
          "description",
          "id",
          "location",
          "name",
          "pictures",
          "price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "amenities": {
            "type": "string"
          },
          "pictures": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BookingInfoDTO": {
        "required": [
          "apartmentInfo",
          "checkIn",
          "checkOut",
          "clientName",
          "guests",
          "id",
          "state"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "clientName": {
            "type": "string"
          },
          "checkIn": {
            "type": "string",
            "format": "date"
          },
          "checkOut": {
            "type": "string",
            "format": "date"
          },
          "guests": {
            "type": "integer",
            "format": "int64"
          },
          "state": {
            "type": "string"
          },
          "apartmentInfo": {
            "$ref": "#/components/schemas/ApartmentInfoDTO"
          }
        }
      },
      "AvailablePeriodInfoDTO": {
        "required": [
          "apartId",
          "date"
        ],
        "type": "object",
        "properties": {
          "apartId": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        }
      }
    }
  }
}